util/romcc: Ensure that bit shift is valid
Found-by: Coverity Scan #1287090
Reviewed-on: https://review.coreboot.org/18020
Tested-by: build bot (Jenkins)

+++ util/romcc/romcc.c
@@ -14250,7 +14250,7 @@ static void compute_closure_variables(struct compile_state *state,
 		}
 		/* Find the lowest unused index value */
 		for(index = 0; index < MAX_INDICIES; index++) {
-			if (!(used_indicies & (1 << index))) {
+			if (!(used_indicies & ((uint64_t)1 << index))) {
 				break;
 			}
 		}
